Remove init() function, move regex compilation to is_roman()
When is_roman() is called, it performs a check if the ROMAN_SYNTAX_REGEX
is already compiled, by checking an inner value of the regex_t struct.
If it is 0, then is not compiled. If not, it compiles it. This allows no
initialization from any user of the library, although it requires an
O(1) useless check for any conversion after the first one.

TODO: fix documentation.