Remove struct, use separate values for int and frac part instead
Adapt the parser data and the variables in the value to roman conversion
function. Change the passed parameters of the function from double
to a pair of int part and frac part. Add functions to convert the
parts from and to double. Add function to convert ints (without frac
part) or doubles to roman, which call the function with both parts.

Move some utility functions around the file.