Use stdint.h to define all integers, stdbool.h for booleans
This assures maximum portability and performance as the integer values
are exactly as small as they could be. At the same time the easy usage
of unsigned values makes the values impossible to hold negative values
when they should not (such as the twelfths from 0 to 11).

The stdbool.h was already used by Numerus only internally, while the
outputs of the functions returning booleans were short integers.
Including stdbool.h into the numerus.h header file allows to use this
macros also for the public user of the library. This has been done for
code clarity as "true" and "false" are more understandable than a 0
and 1, while still being a 0 and 1 if needed.